Various improvements to our test suite
I added the integration_test and fulltest targets to the Makefile, so you can do

    make integration_test

to run the Perl integration test suite, and

    make fulltest

to run both the PHP unit tests and the integration tests. Also, a function
called json_ok($url) was implemented in Test::Cashmusic to reduce
duplicate code.