Setting/removing webhooks on list add/edit/delete - Fixes #143
Separated webhooks from doListSync and removed that function from the basic add,
edit, and delete functions for lists. It's just too much to expect it to work
during a casual operation, so instead we'll make it a discreet thing that has to
be initiated by the user intentionally.

Added manageWebhooks function to the PeoplePlant.

That let us set up a more basic webhook logic. Whenever a list is added we add
any webhooks if a connection is present. On edits we remove existing webbooks,
then add appropriate webhooks for the edit. On deletes we remove webhooks prior
to removing the list.