Moving all element files to a single location in the framework
Previously elements had been split — classes as part of the framework and the
add/edit/etc... files all living in the admin tool. This feels slightly more
correct technically, but it means that element authors would need to place files
in two places to get things going. With those files being dependent on one another
that feels sort of stupid, so now it's as simple as making a bundle a la:

MyStupidElement.php (class)
mystupidelement (folder with all the extra definition)

This drastically improves install and will be a better experience for someone
looking to extend things...