Added "scope" to the sessionGet, sessionSet, and sessionClear functions
No longer just for persistent data, there's now a single global variable registered
specifically to track data on a script scope. The use case is the sign-on element —
need to be able to track opening and closing elements since it works as a wrapper...
but any passing of data between elements that doesn't need to hit persistent storage
could make use of this.