Minor refactors, added POST support to CASHSystem::getURLContents()
TODO: Need to set up some kind of basic responder to verify that POST data is
being sent correctly for testing.

Lots of small enhancements to the getURLContents function: consistent UA string
(set to firefox 7 rather than "php.ini" which would probably be blocked more.)
Also setting context for GET/POST and whether we should return false for 303/404
— defaults true.

Eyes on the future, this function becomes a nice, robust wrapper for
file_get_contents and cURL, allowing the system to select the best choce by
what's available on a given server.