Working Browser ID integration for Sign-In
Changes to the validatelogin call in SystemPlant let us push an address
through as verified, checking it's admin state if required and returning
ID, but not forcing a password comparison.

In CASHSystem we added a method to spit out the appropriate JS to include
the Browser ID JS and bind it to a form, which is then automatically submitted
for processing where the assertion is validated, etc.

This method allows us to handle special-case Browser ID integration as well as
standardized as part of the sign-in element — the verified address can be pushed
into the validate login call or an assertion can be passed to be processed as
part of the validation process.

Word.