FROZEN SCHEMA, Y'ALL
This one's going to need a Duke Leto review, but the final additions and changes
to the schema are in. That includes all the commerce tables, the structure to
support private contact data stored that can work alongside user accounts,
system settings and general analytics, login analytics, and more.

The goal is that this structure should last through initial feature completion,
meaning our upgrade path will be fairly clean until we run into incompatible or
radically new functionality. Or until whatever typo I missed is caught.

Also edited the README and the Makefile as the SQLite file was generated completely
by hand, and should be going forward — pushing to native SQLite types is a good
way to double-check types as well as ensure smooth installation.

And the .sql files themselves have been organized and versioned. More docs to
follow soon.