Better routing of requests for MOST Plant requests - Fixes #6
There's still a switch for truly special cases, but nearly all incoming requests
are now routed through a plant's routeBasicRequest() method. It's a huge
improvement over manually defining required parameters, setting their defaults,
and firing off a method.

We map request types to a method name in a routing table (array), but that's to
allow flexibility. Allowed request methods are also defined there so we don't
lose security on that front. The routeBasicRequest() method then uses built-in
Reflection class magic to compare our CASH request parameters against expected
parameters for the target method. Defaults are respected unless overridden in
the CASH request, and missing required parameters will cause a failure.

As of now the SystemPlant has been ported to the new format, but all other
Plants will be adapted and tests written.