Removing PERL API integration tests
Much of the functionality was duplicated or even expanded in the PHP test suite.
To make sure we don't lose any coverage I added tests validating JSON for every
API request in the general and MC-specific API tests, and added tests for proper
handling of bad API keys in the MC processwebhook stuff.