Digital-only purchase flow for PayPal
End-to-end digital purchase element. It only touches on the functionality in the
core and is artificially restricted to a single item only. Final changes to the
PaypalSeed and tweaks to People/Commerce/SystemPlants to make sure the workflow
is smooth.

Also added a "sandbox" demo for a quick/dirty page that can accept any element.