Migrating elements to new format
All public-facing aspects have been wrapped with mustache templates and the
internal structure of the ElementBase class has been improved dramatically. All
elements now have public "getMarkup" and "getData" functions — getData()
returning all data needed to render the element in an associative array, as
well as the mustache template.

The long-term significance of this is that all elements will be accessible
directly and over an API — so this paves the way for JavaScript includes.

And other shit. (Snuck in markdown on the static content element...)