Consistent asset downloading
First went and added a 'getfulfillmentassets' method to AssetPlant that will
intelligently generate an array of all assets that should be fulfilled for a single
asset. That means files give a single array with their details, and releases list
out an array of all fulfillment assets.

Using that method in the admin, and also updated the following elements:

 - Digital Purchase
 - Download Codes
 - Email Collection
 - Secure Download

They all use the same basic retreval code (minor option var name differences) and
identical markup. That means that they'll all work equally well with a single file
or a complex release. Rad.

Removed all references to 'folder' assets. Because that was really dumb.

Fixes #180