Basic uploading working
Added an assets/uploadform route which displays the most basic filepicker and
upload form. It should be called with the ID of an S3 conncetion, like so:

assets/uploadform/1

Anything else will yield an error message to "please choose connection" and that's
it. The controller gets all the necessary parameters and creates a form that will
allow direct/valid uploads to s3.

The JavaScript stores the filename and key to get the final key on success, which
we can then use to fill out the main location input on the asset form. (This will
need to be a separate form, which is lame, but we can simply do a data_only ajax
request for assets/uploadform/{connection_id} and insert it directly into the html
for the assets edit page.)

Still not setting MIME types. I think we can do that with a copy-in-place request,
so we might want to add something like an ajax/finalizeupload endpoint that just
sets the proper Content-Type to be sure.