Proper password hashing
We first cascade through blowfish, sha512, and sha256 using crypt with a unique
salt on the encryption and system salt added to the password.

If CRYPT_BLOWFISH, CRYPT_SHA512, and CRYPT_SHA256 are all not present we fall
back to a more robust version of the previous hash_hmac implementation where we
use an sha256 encryption on an md5 of the password + system salt, prepending the
key to the final hash so we can get the benefits of strong encryption, unique
keys, and system hash stored outside the db.

Fixes #212