System settings stored in an environment variable first
Basically we're storing a cashmusic_platform_settings ENV variable — if it's found
we never even look at the disk for a cashmusic.ini.php. If it's not found, we read
the cashmusic.ini.php and store it as an ENV variable with putenv, meaning we only
look to the disk once for every run.

But the obvious benefit is we're good for AppFog, load balancing, and beyond.