Thank you to everyone who sponsored a file during our Kickstarter
As we near early public testing of this platform it makes me immensely happy to
push a commit dedicated purely to thanking the people who chose to sponsor a file
in this project. Without all of you and all of the other Kickstarter supporters
we wouldn't have made it this far. So from the bottom of my heart: thank you.

[ci skip]