Pretty major efficiency improvements in the admin
Reduced the number of overall CASHRequest objects per page from the 7-15 range
down to a single instance on most pages, 2 on element pages. This should be a
fairly drastic performance boost at scale. The only major changes to the core
are:

 - Adding support for a public processRequest method to the CASHRequest class
 - Allowing CASHSystem::StartUp(true) to return the initial CASHRequest object

Those two changes allow for the reduction by eliminating the need for a new
CASHRequest object for each request, and by allowing us to recycle the initial
startup object. The admin now runs off a single instance and the same pattern
can be used for any apps.