Simple Mode Mechanism
Did a lot of work on the main controller, streamlining and adding support for a
simple mode. Simple mode itself is set in the theme, so it can change with the
layout of the admin. Notice the new simple.mustache template and the addition of
a "logic" folder to the theme directory.

The template and simplemode.php files are optional, and both must be present in
order for simple mode to be ebabled. Even enabled, a user must have a setting
present (and not false) in their admin settings to see simple mode.

Word.

- Fixes #276