Full translation (#1317)
I fully translated all of the files that are meant to be translated from the tutorial page. I would appreaciate if someone would have a look at these and maybe merge it. I am doing this for my project at university and the deadline is creeping up slowly. Hopefully my work will be of some worth to the main project.